# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnAI - AI-Powered Adaptive Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{conceptId}/review:
    post:
      operationId: submitConceptReview
      tags: [tasks]
      summary: Submit a spaced-repetition recall rating for a concept
      parameters:
        - name: conceptId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConceptReviewInput"
      responses:
        "200":
          description: Updated review schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConceptReviewResult"
        "404":
          description: Concept or schedule not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── NOTE ──────────────────────────────────────────────────────────────────
...
    ConceptReviewInput:
      type: object
      properties:
        rating:
          type: string
          enum: [again, hard, good, easy]
      required:
        - rating
    ConceptReviewResult:
      type: object
      properties:
        conceptId:
          type: integer
        rating:
          type: string
        dueAt:
          type: string
          format: date-time
        stabilityDays:
          type: number
      required:
        - conceptId
        - rating
        - dueAt
        - stabilityDays
...